// src/app/model/response/leave-apply-response.ts
export interface LeaveApplyResponse {
  studentFirstName: null | string;
  studentLastName: null | string;
  fromDate: string;
  toDate: string;
  fromTime: null | string;
  toTime: null | string;
  comments: string;
  leaveStatus: string;
  leaveReason: string;
  leaveType: string;
  applyDate: string;
}